requirements.txt
Flask==3.0.3
lxml==5.3.0
requests==2.32.3
reportlab==4.2.2
python-docx==1.1.2
python-pptx==0.6.23 
html2text==2024.2.26 